fix(01-01): fold rename/widen/deleted_at into the existing unpushed migrations instead of a third file
Per user correction on the Task 2 checkpoint: rewrite migration history directly rather
than stacking a third migration, since both prior migrations are still unpushed/local.

- Delete prisma/migrations/20260707084054_rename_and_widen_existing_tables/ (never applied
  to the live DB — created with --create-only only, safe to remove)
- Hand-edit prisma/migrations/20260616152601_init/migration.sql: CREATE TABLE now emits
  `users` directly (not `User`), id BIGINT AUTO_INCREMENT, deleted_at DATETIME(3) NULL
  column, and the unique index renamed to users_email_key
- Hand-edit prisma/migrations/20260617121303_add_refresh_tokens/migration.sql: id and
  user_id are now BIGINT in the CREATE TABLE block (AUTO_INCREMENT / NOT NULL retained)
- schema.prisma is unchanged — it already reflected this end state from Task 1
- Nothing applied to the live DB; live DB and _prisma_migrations still reflect the OLD
  (User/INTEGER) history — reconciliation DDL + checksum patch to be approved via the next
  checkpoint before Task 3 touches the database

+++ prisma/migrations/20260617121303_add_refresh_tokens/migration.sql
@@ -1,8 +1,8 @@
 -- CreateTable
 CREATE TABLE `refresh_tokens` (
-    `id` INTEGER NOT NULL AUTO_INCREMENT,
+    `id` BIGINT NOT NULL AUTO_INCREMENT,
     `token_hash` VARCHAR(191) NOT NULL,
-    `user_id` INTEGER NOT NULL,
+    `user_id` BIGINT NOT NULL,
     `expires_at` DATETIME(3) NOT NULL,
     `revoked_at` DATETIME(3) NULL,
     `created_at` DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
